pyproject.toml
[project]
name = "buddy-bloom"
version = "0.1.0"
description = "A social networking application using Neo4j and Python"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "pydantic[email]>=2.0",
    "neo4j>=5.0",
    "python-dotenv>=1.0",
    "bcrypt>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "config"]

[tool.uv]
preview = true
